requirements.txt
numpy>=1.22.0
Pillow>=9.0.0
aiohttp>=3.8.0
